Fixing migrations order compile

+++ database/migrations/2014_10_12_000011_create_users_table.php
@@ -22,7 +22,7 @@ class CreateUsersTable extends Migration
             $table->string('email')->unique();
             $table->timestamp('email_verified_at')->nullable();
             $table->string('password');
-            $table->unsignedInteger('address_id');
+            $table->unsignedBigInteger('address_id');
             $table->rememberToken();
             $table->timestamps();
